+++ app/forms/transfer/task/sponsored_support_grant_task_form.rb
@@ -1,0 +1,7 @@
+class Transfer::Task::SponsoredSupportGrantTaskForm < BaseOptionalTaskForm
+  attribute :type, :string
+
+  def type_options
+    Transfer::TasksData.sponsored_support_grant_types.values
+  end
+end