+++ app/models/transfer/tasks_data.rb
@@ -6,4 +6,12 @@ class Transfer::TasksData < ActiveRecord::Base
   def self.policy_class
     TaskListPolicy
   end
+
+  enum :sponsored_support_grant_type,
+    {
+      standard: "standard",
+      fast_track: "fast_track",
+      intermediate: "intermediate",
+      full_sponsored: "full_sponsored"
+    }
 end