+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show `Form a MAT` or `single transfer` in the exports for Transfer projects
 - Add button to allow users to create form a MAT transfer projects
 - Show the date a transfer happened in the Grant management export
+- The Confirm transfer grant funding level task has been added to transfer
+  project task lists
 
 ## [Release-56][release-56]